Update connect.sh
Version 1.1

+++ connect.sh
@@ -11,6 +11,7 @@ path_htb="$HOME/.connect/path_htb.txt"
 path_sp="$HOME/.connect/path_sp.txt"
 path_thm="$HOME/.connect/path_thm.txt"
 path_release_arena="$HOME/.connect/path_release_arena.txt"
+path_desec="$HOME/.connect/path_desec.txt"
 case "$1" in   	                           # Checking options
   -h | --htb)
 	if [[ -e $path_htb ]]	           # Checking file to connect in the VPN   
@@ -18,7 +19,12 @@ case "$1" in   	                           # Checking options
 		cat $path_htb | sh 	   # Executing the file that contain the command
 	else
                 read -p "Path of your HackTheBox VPN file: " path  
-		mkdir ~/.connect
+		if [[ -e ~/.connect ]]
+                then
+                	echo "Connecting..."
+                else
+                	mkdir ~/.connect
+                fi
 		echo "sudo openvpn $path" > $path_htb
 		cat $path_htb | sh
 	fi
@@ -30,7 +36,12 @@ case "$1" in   	                           # Checking options
                 cat $path_sp | sh
         else
                 read -p "Path of your StartPoint VPN file: " path
-		mkdir ~/.connect
+		if [[ -e ~/.connect ]]
+                then
+                	echo "Connecting..."
+                else
+                	mkdir ~/.connect
+                fi
                 echo "sudo openvpn $path" > $path_sp
                 cat $path_sp | sh
         fi
@@ -42,7 +53,12 @@ case "$1" in   	                           # Checking options
                 cat $path_thm | sh
         else
                 read -p "Path of your TryHackMe VPN file: " path
-		mkdir ~/.connect
+		if [[ -e ~/.connect ]]
+                then
+                	echo "Connecting..."
+                else
+                	mkdir ~/.connect
+                fi
 		echo "sudo openvpn $path" > $path_thm
                 cat $path_thm | sh
         fi
@@ -54,7 +70,12 @@ case "$1" in   	                           # Checking options
                 cat $path_release_arena | sh
         else
                 read -p "Path of your ReleaseArenaHTB VPN file: " path
-                mkdir ~/.connect
+                if [[ -e ~/.connect ]]
+                then
+                	echo "Connecting..."
+                else
+                	mkdir ~/.connect
+                fi
                 echo "sudo openvpn $path" > $path_release_arena
                 cat $path_release_arena | sh
 	fi
@@ -64,12 +85,56 @@ case "$1" in   	                           # Checking options
 	echo "Connect version 1.0"
 	exit 0
 	;;
+	-R | --reset-vpn)
+	echo "1 - HTB"
+	echo "2 - StartPointHTB"
+	echo "3 - TryHackMe"
+	echo "4 - ReleaseArenaHTB"
+	read -p "VPN file to remove: " file
+	if [[ $file == 1 ]]
+	then
+		if [[ -e ~/.connect/path_htb.txt ]]
+		then
+			rm ~/.connect/path_htb.txt
+		else
+			echo "File not found"
+		fi
+	elif [[ $file == 2 ]]
+	then
+		if [[ -e ~/.connect/path_sp.txt ]]
+		then
+			rm ~/.connect/path_sp.txt
+		else
+			echo "File not found"
+		fi
+	elif [[ $file == 3 ]]
+	then
+		if [[ -e ~/.connect/path_thm.txt ]]
+		then
+			rm ~/.connect/path_thm.txt
+		else
+			echo "File not found"
+		fi
+	elif [[ $file == 4 ]]
+	then
+		if [[ -e ~/.connect/path_release_arena.txt ]]
+		then
+			rm ~/.connect/path_release_arena.txt
+		else
+			echo "File not found"
+		fi
+	else
+		echo "Invalid option!"
+	fi
+	exit 0
+	;;
 	-help | --help)
 	echo "Usage: $0 -h
 [-h | --htb]   Connect to HackTheBox VPN
 [-s | --start] Connect to StartPointHTB VPN
 [-t | --thm]   Connect to TryHackMe VPN
 [-r | --release-arena] Connect to ReleaseArenaHTB VPN
+[-R | --reset-vpn] Remove a configuration file and reset it to connect to a new VPN file
 [-v | --version] Display the version of your connect file
 [-help | --help] Display this message"
 	exit 0
@@ -87,5 +152,6 @@ echo "Usage: $0 -h
 [-s | --start] Connect to StartPointHTB VPN
 [-t | --thm]   Connect to TryHackMe VPN
 [-r | --release-arena] Connect to ReleaseArenaHTB VPN
+[-R | --reset-vpn] Remove a configuration file and reset it to connect to a new VPN file
 [-v | --version] Display the version of your connect file
 [-help | --help] Display this message"
